config: add new 'format-code-in-docstrings' knob
This commit does the plumbing to make a new formatting option,
'format-code-in-docstrings', available in the configuration for end
users. It is disabled by default (opt-in). It is opt-in at least
initially to reflect a conservative posture. The intent is to make it
opt-out at some point in the future.

The name was inspired by the name of a similar configuration in
rustfmt[1].

[1]: https://rust-lang.github.io/rustfmt/?version=v1.6.0&search=#format_code_in_doc_comments