New AST node for f-string element
This commit adds the new `FStringElement` enum which is either a literal
element or an expression element. To give context, an f-string such as
`f"foo {x}"` is made up of the literal ("foo ") and an expression (x)
elements.

The change here is to have a dedicated literal node for a string inside
an f-string. Earlier, the existing `StringLiteral` node was being used
to represent this. The problem with that is the f-string range doesn't
include the quote while any other string literal would include the
quotes in the range. This creates a problem in the formatter when the
f-string formatting is implemented.